lifted itinerary list state to app.jsx, itinerary module refactoring in progress
