+++ client/components/viewTrip.jsx
@@ -149,8 +149,8 @@ export default class ViewTrip extends React.Component {
     }
     return (
       <>
-        <div className="main trip">
-          <div className="name trip-title">
+        <div className='trip-list-module'>
+          <div className='trip-list-header'>
             <a className={backClass} onClick={() => history.back()}><i className="fas fa-arrow-left back-arrow"></i></a>
             <h2 style={{ fontSize: '1.5rem' }}>{tripName}</h2>
             {editIcon}
